Fixed DrinkCard and Drinks page
