+++ src/pages/DrinkDetails.js
@@ -1,5 +1,5 @@
 import React from 'react';
 
-const DrinkDetails = () => <h1>Page</h1>;
+const DrinkDetails = () => <h1>Detalhes do Drink</h1>;
 
 export default DrinkDetails;
